+++ packages/core/lib/combo/patch-build/group.ts
@@ -52,6 +52,7 @@ export const PATCH_GROUPS = [
   'OOT_FAST_BOAT',
   'MM_WELL_OPEN_OR_REMORSELESS',
   'MM_JP_LAYOUT_DEKU_PALACE',
+  'MM_JP_LAYOUT_STONE_TOWER',
 ] as const;
 
 export type PatchGroup = typeof PATCH_GROUPS[number];